# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "register user and creats a acoount i the database",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstname": {
                  "example": "any"
                },
                "lastname": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "adress": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200, 500": {
            "description": ""
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "login user",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "200, 500": {
            "description": ""
          }
        }
      }
    },
...
    "/products/search/brand/{brandId}": {
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Gets the list of all products by brand.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "brandId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200, 500": {
            "description": ""
          }
        }
      }
    },
...
    "/cart/add/{productId}/{units}": {
      "post": {
        "tags": [
          "Cart"
        ],
        "description": "Adds a product to a cart",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "units",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/cart/checkout/now/{cartId}/{userId}": {
      "post": {
        "tags": [
          "Cart"
        ],
        "description": "checkout cart and creates a order.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
...
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "description": "Gets the list of all available brands.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/brands/add": {
      "post": {
        "tags": [
          "Brands"
        ],
        "description": "Creates a new brand.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Brand"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/brands/edit/{brandId}": {
      "put": {
        "tags": [
          "Brands"
        ],
        "description": "Edits an existing brand.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "brandId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/brands/delete/{brandId}": {
      "delete": {
        "tags": [
          "Brands"
        ],
        "description": "Deletes a brand.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "brandId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
...
  },
  "definitions": {
    "Brand": {
      "type": "object",
      "properties": {
        "Name": {
          "type": "string",
          "example": "Apple"
        }
      },
      "required": [
        "Name"
      ]
    },
...
  }
}